feat(user-ui): change models
