+++ packages/models/src/deliberation_contents/deliberation_content.rs
@@ -5,7 +5,7 @@ use validator::Validate;
 use crate::deliberation_user::DeliberationUser;
 use crate::elearnings::elearning::ElearningCreateRequest;
 use crate::elearnings::elearning::{self, Elearning};
-use crate::Question;
+use crate::{Question, ResourceFile};
 
 #[derive(Validate)]
 #[api_model(base = "/v2/deliberations/:deliberation-id/contents", table = deliberation_contents, action = [create(users = Vec<i64>, elearnings = Vec<ElearningCreateRequest>)])]
@@ -43,4 +43,8 @@ pub struct DeliberationContent {
     #[api_model(summary, action = create, type = JSONB, version = v0.1, action_by_id = update)]
     #[serde(default)]
     pub questions: Vec<Question>,
+
+    #[api_model(summary, many_to_many = deliberation_study_materials, table_name = resources, foreign_primary_key = resource_id, foreign_reference_key = deliberation_id)]
+    #[serde(default)]
+    pub resources: Vec<ResourceFile>,
 }
