+++ packages/user-ui/src/pages/projects/_id/components/final_draft/i18n.rs
@@ -1,0 +1,43 @@
+use bdk::prelude::*;
+
+translate! {
+    FinalDraftTranslate;
+
+    title: {
+        ko: "최종 권고안",
+        en: "Final Recommendation",
+    },
+    necessary: {
+        ko: "[필수]",
+        en: "[Necessary]"
+    }
+    plural: {
+        ko: "[복수]",
+        en: "[Plural]"
+    }
+    unit: {
+        ko: "명",
+        en: "Unit"
+    }
+    subjective_answer: {
+        ko: "주관식 답변",
+        en: "Subjective Answer"
+    }
+    update: {
+        ko: "수정하기",
+        en: "Update"
+    }
+
+    name: {
+        ko: "제목",
+        en: "Title"
+    }
+    description: {
+        ko: "내용",
+        en: "Description"
+    }
+    name_hint: {
+        ko: "제목 입력",
+        en: "Input Title"
+    }
+}
