+++ packages/user-ui/src/pages/projects/_id/components/final_draft/mod.rs
@@ -1,0 +1,5 @@
+mod controllers;
+mod final_draft;
+mod i18n;
+
+pub use final_draft::FinalDraft;
