+++ packages/user-ui/src/pages/projects/_id/components/final_survey/i18n.rs
@@ -1,0 +1,63 @@
+use bdk::prelude::*;
+
+translate! {
+    FinalVoteModalTranslate;
+
+    title: {
+        ko: "투표 전, 다시 한번 확인해주세요.",
+        en: "Please check again before voting."
+    }
+    description: {
+        ko: "투표는 익명으로 진행되며, 한 번 제출한 투표는 변경할 수 없습니다. \n이는 투표의 공정성과 보안을 유지하기 위한 조치입니다. \n투표 후에는 수정이 불가능하며, 결과의 신뢰성을 위해 이 점을 양해 부탁드립니다.",
+        en: "Voting is anonymous and once a vote is submitted, it cannot be changed. \nThis is a measure to maintain the fairness and security of the vote. \nAfter voting, it cannot be changed, so please understand this for the reliability of the results."
+    }
+    complete_voting: {
+        ko: "투표 완료하기",
+        en: "Complete Voting"
+    }
+    cancel: {
+        ko: "취소",
+        en: "Cancel"
+    }
+}
+
+translate! {
+    FinalSurveyTranslate;
+
+    title: {
+        ko: "최종 투표 주제",
+        en: "Final Vote Topic",
+    }
+    see_detail: {
+        ko: "자세히 보기",
+        en: "See Detail"
+    }
+    my_answer: {
+        ko: "내가 작성한 답변",
+        en: "My Answer"
+    }
+    response_per_question: {
+        ko: "질문별 응답",
+        en: "Responses to each question"
+    }
+    necessary: {
+        ko: "[필수]",
+        en: "[Necessary]"
+    }
+    plural: {
+        ko: "[복수]",
+        en: "[Plural]"
+    }
+    unit: {
+        ko: "명",
+        en: "Unit"
+    }
+    subjective_answer: {
+        ko: "주관식 답변",
+        en: "Subjective Answer"
+    }
+    submit: {
+        ko: "제출하기",
+        en: "Submit"
+    }
+}
