+++ packages/user-ui/src/pages/projects/_id/components/final_survey/mod.rs
@@ -1,0 +1,12 @@
+mod controllers;
+mod i18n;
+
+mod final_statistics;
+mod final_survey_info;
+mod final_survey_question;
+mod final_vote_modal;
+mod my_final_survey;
+
+mod final_survey;
+
+pub use final_survey::FinalSurvey;
