+++ packages/user-ui/src/pages/projects/_id/components/mod.rs
@@ -3,16 +3,13 @@ pub mod comment_item;
 pub mod comments;
 pub mod deliberation;
 pub mod discussion;
+
 pub mod final_draft;
-pub mod final_statistics;
+
 pub mod final_survey;
-pub mod final_survey_info;
-pub mod final_survey_question;
-pub mod final_vote_modal;
+
 pub mod i18n;
 pub mod multiple_objective;
-pub mod my_final_survey;
-pub mod not_complete_survey_modal;
 pub mod project_header;
 pub mod sample_survey;
 pub mod single_objective;
